+++ src/SOAPify/SOAPClassify.py
@@ -17,14 +17,22 @@ class SOAPclassification:
 
 @dataclass
 class SOAPReferences:
-    """Utility class to store the information about the SOAP classification of a system."""
+    """
+    Utility class to store the spectra selected for a environments dictionary.
 
-    names: "list[str]"  #:: stores the names of the references
-    spectra: "np.ndarray[float]"  #:: stores the SOAP vector of the references
+    """
+
+    names: "list[str]"  #: stores the names of the references
+    spectra: "np.ndarray[float]"  #: stores the SOAP vector of the references
     lmax: int
     nmax: int
 
     def __len__(self) -> int:
+        """returns the lenght of the dictionary, aka the number of stored spectra
+
+        Returns:
+            int: the number of stored spectra
+        """
         return len(self.names)
 
 
@@ -104,13 +112,20 @@ def loadRefs(
     return spectra, legend
 
 
-# TODO: create a version that acceps and arbitray number of datasets
 def mergeReferences(*x: SOAPReferences) -> SOAPReferences:
+    """Merges a list of `SOAPReferences` into a single object
+
+    Raises:
+        ValueError: if the lmax and the nmax of the references are not the same
+
+    Returns:
+        SOAPReferences: a new `SOAPReferences` that contains the concatenated list of references
+    """
     names = []
     for i in x:
         names += i.names
         if x[0].nmax != i.nmax or x[0].nmax != i.lmax:
-            raise Exception("nmax or lmax are not the same in the two references")
+            raise ValueError("nmax or lmax are not the same in the two references")
     return SOAPReferences(
         names,
         np.concatenate([i.spectra for i in x]),
@@ -122,6 +137,13 @@ def mergeReferences(*x: SOAPReferences) -> SOAPReferences:
 def saveReferences(
     h5position: "h5py.Group|h5py.File", targetDatasetName: str, refs: SOAPReferences
 ):
+    """Export the given references in the indicated group/hdf5 file
+
+    Args:
+        h5position (h5py.Group|h5py.File): The file object of the group where to save the references
+        targetDatasetName (str): the name to give to the list of references
+        refs (SOAPReferences): the `SOAPReferences` object to be exported
+    """
     whereToSave = h5position.require_dataset(
         targetDatasetName,
         shape=refs.spectra.shape,
@@ -135,7 +157,17 @@ def saveReferences(
     whereToSave.attrs.create("names", refs.names)
 
 
-def getReferencesFromDataset(dataset: h5py.Dataset):
+def getReferencesFromDataset(dataset: h5py.Dataset) -> SOAPReferences:
+    """Given a `h5py.Dataset` returns a `SOAPReferences` with the initializated data
+
+        TODO: check if the dataset contains the needed references
+
+    Args:
+        dataset (h5py.Dataset): the dataset with the references
+
+    Returns:
+        SOAPReferences: the prepared references container
+    """
     fingerprints = dataset[:]
     names = dataset.attrs["names"].tolist()
     lmax = dataset.attrs["lmax"]
@@ -184,6 +216,18 @@ def createReferencesFromTrajectory(
 def getDistanceBetween(
     data: np.ndarray, spectra: np.ndarray, distanceCalculator: Callable
 ) -> np.ndarray:
+    """Generate an array with the distances between the the data and the given collection of `spectra`
+
+        TODO: enforce the np.ndarray
+
+    Args:
+        data (np.ndarray): the array of the data
+        spectra (np.ndarray): the references
+        distanceCalculator (Callable): the function to calculate the distances
+
+    Returns:
+        np.ndarray: the array of the distances (the shape is `(data.shape[0], spectra.shape[0])`)
+    """
     toret = np.zeros((data.shape[0], spectra.shape[0]), dtype=data.dtype)
     for j in range(spectra.shape[0]):
         for i in range(data.shape[0]):
@@ -196,9 +240,21 @@ def getDistancesFromRef(
     references: SOAPReferences,
     distanceCalculator: Callable,
     doNormalize: bool = False,
-):
+) -> np.ndarray:
+    """generates the distances between a SOAP-hdf5 trajectory and the given references
+
+    Args:
+        SOAPTrajData (h5py.Dataset): the dataset containing the SOAP trajectory
+        references (SOAPReferences): the contatiner of the references
+        distanceCalculator (Callable): the function to calculate the distances
+        doNormalize (bool, optional): informs the function if the given data needs to be normalized before caclulating the distanceis already normalized. Defaults to False.. Defaults to False.
+
+    Returns:
+        np.ndarray: the "trajectory" of distance from the given references
+    """
     # TODO use the dataset chunking
     CHUNK = 100
+    # =min(100,SOAPTrajData.chunks[0])
     # assuming shape is (nframes, natoms, nsoap)
     currentFrame = 0
     doconversion = SOAPTrajData.shape[-1] != references.spectra.shape[-1]
@@ -224,6 +280,7 @@ def getDistancesFromRef(
 def getDistancesFromRefNormalized(
     SOAPTrajData: h5py.Dataset, references: SOAPReferences
 ):
+    """shortcut for `SOAPify.SOAPClassify.getDistancesFromRef(SOAPTrajData,references,True)`, see :func:`SOAPify.SOAPClassify.getDistancesFromRef`"""
     return getDistancesFromRef(
         SOAPTrajData, references, SOAPdistanceNormalized, doNormalize=True
     )
@@ -235,6 +292,16 @@ def classify(
     distanceCalculator: Callable,
     doNormalize: bool = False,
 ) -> SOAPclassification:
+    """generates the distances from the given references and then classyfy all of the atoms by the colosest element in the dictionary
+
+    Args:
+        SOAPTrajData (h5py.Dataset): the dataset containing the SOAP trajectory
+        references (SOAPReferences):  the contatiner of the references
+        distanceCalculator (Callable): the function to calculate the distances
+        doNormalize (bool, optional): informs the function if the given data needs to be normalized before caclulating the distanceis already normalized. Defaults to False.. Defaults to False.
+    Returns:
+        SOAPclassification: The result of the classification
+    """
     info = getDistancesFromRef(
         SOAPTrajData, references, distanceCalculator, doNormalize
     )